pyproject.toml
[project]
name = "seo-blog-agent"
version = "0.1.0"
description = "SEO-optimized blog writer agent using LangGraph and CrewAI"
requires-python = ">=3.11"
dependencies = [
    "crewai[tools]>=0.100.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-groq>=0.2.0",
    "langchain-mistralai>=0.2.0",
    "google-search-results>=2.4.2",
    "streamlit>=1.40.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.5.0",
    "textstat>=0.7.4",
    "jinja2>=3.1.0",
    "python-frontmatter>=1.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
